+++ services/api/carlos/api/routes/device_server_routes/__init__.py
@@ -1,0 +1,3 @@
+__all__ = ["device_server_router"]
+
+from .device_server_routes import device_server_router
