+++ services/api/carlos/api/routes/device_server_routes/state.py
@@ -1,0 +1,6 @@
+__all__ = ["DEVICE_CONNECTION_MANAGER"]
+
+from carlos.edge.server.connection import DeviceConnectionManager
+
+DEVICE_CONNECTION_MANAGER = DeviceConnectionManager()
+"""Singleton instance of the DeviceConnectionManager."""
